Fix 'array subscript is char' warning on NetBSD
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@32113 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/cartconv.c
@@ -483,7 +483,7 @@ static int checkflag(char *flg, char *arg)
 {
     int i;
 
-    switch (tolower(flg[1])) {
+    switch (tolower((int)(flg[1]))) {
         case 'f':
             printinfo(arg);
             return 2;
